+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Session` class `to_dataframe` method
 ### Changed
 - `LLMModel` enum updated
 - Test system modified
+- `README.md` updated
 ## [1.0] - 2025-11-17
 ### Added
 - `Response` class `xml_tree` property